fixed initialization for union punning

+++ src/build_hash.cpp
@@ -76,8 +76,8 @@ int main(int argc, char** argv) {
     }
 
     float unk_prob = global::default_unk_prob;
-    uint8_t probs_quantization_bits = global::default_probs_quantization_bits;
-    uint8_t backoffs_quantization_bits =
+    uint32_t probs_quantization_bits = global::default_probs_quantization_bits;
+    uint32_t backoffs_quantization_bits =
         global::default_backoffs_quantization_bits;
     const char* input_dir = ".";
     const char* arpa_filename = nullptr;
@@ -90,10 +90,10 @@ int main(int argc, char** argv) {
                   << std::endl;
     }
     if (parser.parsed("p")) {
-        probs_quantization_bits = parser.get<uint8_t>("p");
+        probs_quantization_bits = parser.get<uint32_t>("p");
     }
     if (parser.parsed("b")) {
-        backoffs_quantization_bits = parser.get<uint8_t>("b");
+        backoffs_quantization_bits = parser.get<uint32_t>("b");
     }
 
     auto dir = parser.get<std::string>("dir");
